requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
oauthlib==3.0.1
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
urllib3==1.24.1
Werkzeug==0.15.1
